start of a cdf specific wrapfile

+++ jobsub_tools/prd/jobsub_tools/v1_3/Linux-2/test/jobsub_ups_unit_test.sh
@@ -4,6 +4,7 @@ source $UPS_DIR/test/unittest.bash
 
 test_setup() {
    testdir=/tmp/t$$
+   testout=/tmp/test_case_out_$$
    mkdir -p $testdir/tmp
    mkdir -p $testdir/exec
    export CONDOR_TMP=$testdir/tmp
@@ -11,8 +12,10 @@ test_setup() {
 }
 
 test_teardown() {
-   return 0
-   rm -rf $testdir
+   if [ "$SAVE_TEST_OUTPUT" = "" ]; then
+       rm -rf $testdir
+       rm -rf $testout
+   fi
 }
 
 print_cmd_file() {
@@ -92,6 +95,7 @@ test_OS() {
    # succeeds if foo and bar are in the DesiredOS line
    grep 'DesiredOS *= *"foo,bar"' $file /dev/null && grep 'requirements.*DesiredOS' $file /dev/null
 }  
+
 
 
 testsuite setups_suite	\
